requirements.txt
feedparser==6.0.11
python-dateutil==2.9.0.post0
beautifulsoup4==4.12.3
lxml==5.2.2
backports.zoneinfo; python_version < "3.9"
